Renamed Styled to styled
