+++ src/client/src/shell/notification/styled.tsx
